pyproject.toml
# Modified: 2026-02-08T18:00:00Z | Author: COPILOT | Change: Add SK extras (ollama,hugging_face,chroma), security optional deps (msticpy, azure-mgmt-security), sentence-transformers
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slate"
version = "2.4.0"
description = "System Learning Agent for Task Execution"
readme = "README.md"
license = {text = "EOSL-1.0"}
requires-python = ">=3.11"
authors = [
    {name = "SynchronizedLivingArchitecture"}
]
keywords = ["ai", "agent", "orchestration", "pytorch", "gpu"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "psutil>=5.9.0,<8",
    "rich>=13.0.0,<15",
    "aiohttp>=3.9.0,<4",
    "python-dotenv>=1.0.0,<2",
    "fastapi>=0.109.0,<1",
    "uvicorn[standard]>=0.27.0,<1",
    "websockets>=12.0,<17",
    "httpx>=0.26.0,<1",
    "chromadb>=0.4.22,<2",
    "typing-extensions>=4.0.0,<5",
    "watchfiles>=1.0.0,<2",
    "semantic-kernel[ollama,hugging_face,chroma]>=1.39.0,<2",
    "github-copilot-sdk>=0.1.0",
    "pydantic>=2.0,<3",
]

[project.optional-dependencies]
ai = [
    "transformers>=4.40.0,<6",
    "accelerate>=0.30.0,<2",
    "peft>=0.10.0,<1",
    "sentence-transformers>=5.0.0,<6",
]
security = [
    "msticpy>=2.18.0,<3",
    "azure-mgmt-security>=7.0.0,<8",
]
dev = [
    "pytest>=8.0.0,<10",
    "pytest-asyncio>=0.23.0,<2",
    "pytest-cov>=4.1.0,<8",
    "ruff>=0.2.0,<1",
    "mypy>=1.8.0,<2",
]
observability = [
    "opentelemetry-api>=1.20.0,<2",
    "opentelemetry-sdk>=1.20.0,<2",
    "opentelemetry-exporter-otlp>=1.20.0,<2",
]

[project.scripts]
slate-status = "slate.slate_status:main"
slate-runtime = "slate.slate_runtime:main"
slate-benchmark = "slate.slate_benchmark:main"
slate-hardware = "slate.slate_hardware_optimizer:main"
slate-copilot = "slate.slate_copilot_sdk:main"
slate-copilot-install = "slate.install_copilot_sdk:main"

[project.urls]
Homepage = "https://github.com/SynchronizedLivingArchitecture/S.L.A.T.E"
Repository = "https://github.com/SynchronizedLivingArchitecture/S.L.A.T.E"
Issues = "https://github.com/SynchronizedLivingArchitecture/S.L.A.T.E/issues"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    "actions-runner*",
    "_temp",
    ".venv",
    "__pycache__",
    ".git",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
prerelease = "allow"
